9-24 commit
Minimally viable product with a decent amount of testing. Textboxes that
receive invalid input will simply be cleared - it would be nice to show
an message specifiying where the error came from.

TODO:
- Use drop down menus for selecting credits/grades?
- Show where invalid input occurred
- Class 1 : Grade isn't tab stopping...wtf?
- Edit a class name by tapping the textblock (originally containing
something like "Class 1". Show a textbox for the user to input a new
name, reset the TextBlock.Text to that name.
- Add another classes. We currently show 6 classes and you can leave as
many was you want blank. Wouldn't hurt to adhere to the over-achievers.
- Calculate grades for an individual class. (I.e., HW - 20%, Exam 1 -
25%, Exam 2 - 25%, Final - 30% type thing)