9-26 commit
Finished a bunch of error checking/validation. Still have some written
down in a notebook that are worth looking into. Added small, medium and
wide icons for the start menu as well as a nice splash screen. Looks
like a built in app using transparent icons.

Before submitting it to the store, add the store logo (which is
transparent, may need a background first). I'd like to be able to edit
class names.

Now we've got a semi-well tested minimally viable product + some
asthetic improvements.